+++ admin/tool/muprog/db/messages.php
@@ -29,6 +29,12 @@
 defined('MOODLE_INTERNAL') || die();
 
 $messageproviders = [
+    'cc_supervisor_notification' => [
+        'defaults' => [
+            'popup' => MESSAGE_PERMITTED + MESSAGE_DEFAULT_ENABLED,
+            'email' => MESSAGE_PERMITTED + MESSAGE_DEFAULT_ENABLED,
+        ],
+    ],
     'allocation_notification' => [
         'defaults' => [
             'popup' => MESSAGE_PERMITTED + MESSAGE_DEFAULT_ENABLED,